[#126] fix: useSearchTerm 로직 분리
