+++ client/src/components/productList/SearchBox/index.tsx
@@ -1,15 +1,9 @@
-/* eslint-disable @typescript-eslint/no-use-before-define */
-/* eslint-disable consistent-return */
-/* eslint-disable react/no-array-index-key */
-import { FC, useContext, useEffect, useRef, useState } from 'react';
+import { FC, useContext, useEffect, useRef } from 'react';
 
 import { SearchBoxUnderlineSVG, SearchSVG, XSVG } from '~/assets';
-import { FilterContext } from '~/pages/ProductList';
-import { removeSearchValue, setSearchValue } from '~/stores/productListModule';
-import {
-  getValueOnLocalStorage,
-  setValueOnLocalStorage,
-} from '~/utils/localStorage';
+import { FetchContext, FilterContext } from '~/pages/ProductList';
+import { startFetch } from '~/stores/fetchModule';
+import { removeSearchValue } from '~/stores/productListModule';
 
 import RecentlySearchTermBox from '../RecentlySearchTermBox';
 import {
@@ -19,6 +13,7 @@ import {
   SearchLine,
   ValueRemoveButton,
 } from './index.style';
+import useSearchTerm from './useSearchTerm';
 
 const SearchBox: FC = () => {
   const searchTermRef = useRef<HTMLInputElement>();
@@ -27,6 +22,7 @@ const SearchBox: FC = () => {
     useSearchTerm(searchTermRef);
 
   const { dispatch, ...filterValue } = useContext(FilterContext);
+  const { dispatch: fetchDispatch } = useContext(FetchContext);
   const currentSearchTerm = filterValue.state.search;
 
   useEffect(() => {
@@ -35,6 +31,7 @@ const SearchBox: FC = () => {
 
   const removeTerm = () => {
     dispatch(removeSearchValue());
+    fetchDispatch(startFetch());
     searchTermRef.current.value = '';
   };
 
@@ -63,60 +60,3 @@ const SearchBox: FC = () => {
 };
 
 export default SearchBox;
-
-/**
- * 최근 검색어 리스트를 관리하는 훅입니다.
- */
-const useSearchTerm = (
-  searchTermRef: React.MutableRefObject<HTMLInputElement>,
-) => {
-  const { dispatch } = useContext(FilterContext);
-
-  const [termList, setTermList] = useState<string[]>(
-    getValueOnLocalStorage('recentlySearchTerm'),
-  );
-
-  /**
-   * input에 있는 value를 상태에 반영합니다.
-   * 만약, 빈 값을 받는다면 store의 search 값을 제거하는 요청을 보냅니다.
-   */
-  const handleSearchTrigger = (e: React.FormEvent<HTMLFormElement>): void => {
-    e.preventDefault();
-
-    const term = searchTermRef.current.value;
-    if (term === '') return dispatch(removeSearchValue());
-
-    addTermOnList(term);
-    dispatch(setSearchValue(term));
-  };
-
-  /**
-   * term을 리스트에 추가합니다.
-   * localStorage에 이미 term이 있을 경우, 추가하지 않습니다.
-   */
-  const addTermOnList = (term: string) => {
-    const storedTermValue = getValueOnLocalStorage('recentlySearchTerm');
-    if (storedTermValue && storedTermValue.includes(term)) return;
-
-    const addedTermList = storedTermValue ? [...storedTermValue, term] : [term];
-    reflectRenewTerm(addedTermList);
-  };
-
-  /**
-   * target을 termList에서 제거합니다.
-   */
-  const removeTermOnList = (target: string) => {
-    const removedTermList = termList.filter((term) => term !== target);
-    reflectRenewTerm(removedTermList);
-  };
-
-  /**
-   * localStorage와 termList에 최근 검색어를 갱신합니다.
-   */
-  const reflectRenewTerm = (renewTermList: string[]) => {
-    setValueOnLocalStorage('recentlySearchTerm', renewTermList);
-    setTermList(renewTermList);
-  };
-
-  return { termList, handleSearchTrigger, removeTermOnList };
-};
